Fix SocketFactory creation
try to instantiate socketFactoryClass. If it fails try to use the geDefault method.
This is the way documented in JavaDoc and also used by i.E. angus-mail.

getDefault is the only way to get fallback to "javax.net.ssl.SSLSocketFactory" work. Because
SSLSocketFactory is an abstract class and not instantiable without subclassing.